refactored TransactionInfoBuilder name to TransactionInfoService
