+++ src/main/kotlin/it/pagopa/ecommerce/transactions/scheduler/deadletter/NotificationDeadLetterConsumer.kt
@@ -4,6 +4,7 @@ import com.azure.spring.messaging.AzureHeaders
 import com.azure.spring.messaging.checkpoint.Checkpointer
 import it.pagopa.ecommerce.commons.queues.StrictJsonSerializerProvider
 import it.pagopa.ecommerce.transactions.scheduler.repositories.DeadLetterEventRepository
+import it.pagopa.ecommerce.transactions.scheduler.services.TransactionInfoBuilder
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.integration.annotation.ServiceActivator
